+++ translations/location_history_visualizer_ar.ts
@@ -20,24 +20,12 @@
         <translation>&amp;إعدادات</translation>
     </message>
     <message>
-        <source>Map display...</source>
-        <translation>عرض الخريطة...</translation>
-    </message>
-    <message>
         <source>&amp;Help</source>
         <translation>&amp;مساعدة</translation>
     </message>
     <message>
         <source>About</source>
         <translation>حول</translation>
-    </message>
-    <message>
-        <source>Open JSON...</source>
-        <translation>فتح JSON...</translation>
-    </message>
-    <message>
-        <source>No file loaded</source>
-        <translation>لم يتم تحميل ملف</translation>
     </message>
     <message>
         <source>Counts</source>
@@ -262,9 +250,6 @@
         <source>&lt;b&gt;GitHub repository:&lt;/b&gt; %1</source>
         <translation>&lt;b&gt;مستودع GitHub:&lt;/b&gt; %1</translation>
     </message>
-</context>
-<context>
-    <name>LocationHistory::MapDisplayDialog</name>
     <message>
         <source>Map display</source>
         <translation>عرض الخريطة</translation>