+++ translations/location_history_visualizer_es.ts
@@ -20,24 +20,12 @@
         <translation>&amp;Configuración</translation>
     </message>
     <message>
-        <source>Map display...</source>
-        <translation>Visualización del mapa...</translation>
-    </message>
-    <message>
         <source>&amp;Help</source>
         <translation>A&amp;yuda</translation>
     </message>
     <message>
         <source>About</source>
         <translation>Acerca de</translation>
-    </message>
-    <message>
-        <source>Open JSON...</source>
-        <translation>Abrir JSON...</translation>
-    </message>
-    <message>
-        <source>No file loaded</source>
-        <translation>Ningún archivo cargado</translation>
     </message>
     <message>
         <source>Counts</source>
@@ -262,9 +250,6 @@
         <source>&lt;b&gt;GitHub repository:&lt;/b&gt; %1</source>
         <translation>&lt;b&gt;Repositorio de GitHub:&lt;/b&gt; %1</translation>
     </message>
-</context>
-<context>
-    <name>LocationHistory::MapDisplayDialog</name>
     <message>
         <source>Map display</source>
         <translation>Visualización del mapa</translation>