+++ translations/location_history_visualizer_ru.ts
@@ -20,24 +20,12 @@
         <translation>&amp;Настройки</translation>
     </message>
     <message>
-        <source>Map display...</source>
-        <translation>Отображение карты...</translation>
-    </message>
-    <message>
         <source>&amp;Help</source>
         <translation>&amp;Справка</translation>
     </message>
     <message>
         <source>About</source>
         <translation>О программе</translation>
-    </message>
-    <message>
-        <source>Open JSON...</source>
-        <translation>Открыть JSON...</translation>
-    </message>
-    <message>
-        <source>No file loaded</source>
-        <translation>Файл не загружен</translation>
     </message>
     <message>
         <source>Counts</source>
@@ -262,9 +250,6 @@
         <source>&lt;b&gt;GitHub repository:&lt;/b&gt; %1</source>
         <translation>&lt;b&gt;Репозиторий GitHub:&lt;/b&gt; %1</translation>
     </message>
-</context>
-<context>
-    <name>LocationHistory::MapDisplayDialog</name>
     <message>
         <source>Map display</source>
         <translation>Отображение карты</translation>